refactor: move potential non singletons to utils
