+++ packages/relay/test/login.test.ts
@@ -4,7 +4,7 @@ import { ethers } from 'hardhat'
 import nock from 'nock'
 import { Synchronizer } from '@unirep/core'
 import { TWITTER_CLIENT_ID, TWITTER_CLIENT_KEY } from '../src/config'
-import { TransactionManager } from '../src/services/singletons/TransactionManager'
+import { TransactionManager } from '../src/services/utils/TransactionManager'
 import { userService } from '../src/services/UserService'
 import { UserRegisterStatus } from '../src/types'
 import { CLIENT_URL } from './configs'
